requirements.txt
amqp==2.5.2
asgiref==3.2.3
Babel==2.9.1
billiard==3.6.3.0
celery==4.4.1
Django==2.2.10
gunicorn==20.0.4
importlib-metadata==1.5.0
kombu==4.6.8
mysql-connector==2.2.9
mysqlclient==1.4.6
python-stdnum==1.13
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
suds-jurko==0.6
vatnumber==1.2
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
